Mux now uses an enum; corrected non-muxed signals in muxed messages.
It is possible for multiplexed messages to contain non-muxed signals,
which are always present regardless of the multiplexer value. These
signals were not handled correctly, because the fields CANsignal::mux
and CANsignal::muxNdx did not provide an indication of this. With the
new enum, this is now handled more gracefully.

The mux grammar was also renamed to muxer in an attempt to make the
separation of the mux types more obvious.